+++ test/roots/test-autodoc-emmylua/lua/doctype.lua
@@ -1,0 +1,13 @@
+--- @meta
+
+--- Description
+local doctype = {}
+
+--- Description
+--- @doctype const
+doctype.x = 10
+
+--- Description
+function doctype.foo(x, y, z) end
+
+return doctype